fix: dialog content screen overflow
If the screen gets small enough for the dialog content to overflow it,
the content is not scrollable and the user has no chance to see all of
it. This can easily happen on mobile with keyboard open.

The solution I chose is to limit the height to viewport height and
let the browser handle the overflow by showing a scrollbar.