fix: sheet content vertical overflow
If the sheet content grew too large, the portion overflowing the viewport
would be unreachable as there was no way to scroll to it. This can easily
happen on mobile with keyboard open.

The solution I chose is to limit the height to viewport height and
let the browser handle the overflow by showing a scrollbar.